+++ packages/commons/src/helpers/index.ts
@@ -1,3 +1,5 @@
 export * from './pagination'
 export * from './getUniqueValues'
 export * from './arrayToHash'
+export * from './random'
+export * from './arrayUnique'