+++ packages/gatekeeper-internal/config/default.js
@@ -4,6 +4,7 @@ require('dotenv').config()
 const {
   AUTH_SERVICE_URL,
   GROUPS_SERVICE_URL,
+  POSTS_SERVICE_URL,
   AUTH_SERVICE_KEY,
   GROUPS_SERVICE_KEY,
   TAGS_SERVICE_KEY,
@@ -20,6 +21,7 @@ module.exports = {
   services: {
     auth: `http://${AUTH_SERVICE_URL}`,
     groups: `http://${GROUPS_SERVICE_URL}`,
+    posts: `http://${POSTS_SERVICE_URL}`,
   },
   appKeys: {
     auth: AUTH_SERVICE_KEY,