+++ packages/gatekeeper-public/src/services/groups.ts
@@ -32,6 +32,56 @@ export default {
       restricted: false,
     },
     {
+      path: '/:slug/members/:user',
+      method: http.DELETE,
+      restricted: true,
+    },
+    {
+      path: '/:slug/admins',
+      method: http.GET,
+      restricted: false,
+    },
+    {
+      path: '/:slug/admins',
+      method: http.POST,
+      restricted: true,
+    },
+    {
+      path: '/:slug/admins/:user',
+      method: http.DELETE,
+      restricted: true,
+    },
+    {
+      path: '/:slug/invitations',
+      method: http.POST,
+      restricted: true,
+    },
+    {
+      path: '/:slug/invitations',
+      method: http.GET,
+      restricted: true,
+    },
+    {
+      path: '/:slug/requests',
+      method: http.POST,
+      restricted: true,
+    },
+    {
+      path: '/:slug/requests',
+      method: http.GET,
+      restricted: true,
+    },
+    {
+      path: '/invitations/my',
+      method: http.GET,
+      restricted: true,
+    },
+    {
+      path: '/invitations/:id',
+      method: http.DELETE,
+      restricted: true,
+    },
+    {
       path: '/:slug',
       method: http.GET,
       restricted: false,