+++ packages/client-queue/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.12](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/client-queue@0.6.11...@gtms/client-queue@0.6.12) (2020-06-07)
+
+**Note:** Version bump only for package @gtms/client-queue
+
+
+
+
+
+## [0.6.11](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/client-queue@0.6.10...@gtms/client-queue@0.6.11) (2020-06-07)
+
+**Note:** Version bump only for package @gtms/client-queue
+
+
+
+
+
 ## [0.6.10](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/client-queue@0.6.9...@gtms/client-queue@0.6.10) (2020-05-23)
 
 **Note:** Version bump only for package @gtms/client-queue