+++ packages/lib-api/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/lib-api@0.3.1...@gtms/lib-api@0.3.2) (2020-06-07)
+
+**Note:** Version bump only for package @gtms/lib-api
+
+
+
+
+
+## [0.3.1](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/lib-api@0.3.0...@gtms/lib-api@0.3.1) (2020-06-07)
+
+**Note:** Version bump only for package @gtms/lib-api
+
+
+
+
+
+# [0.3.0](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/lib-api@0.2.9...@gtms/lib-api@0.3.0) (2020-05-26)
+
+
+### Features
+
+* GEOT-266: add favs endpoints ([#98](https://github.com/mariusz-kabala/gtms-backend/issues/98)) ([b296aff](https://github.com/mariusz-kabala/gtms-backend/commit/b296aff2f4b87ad9a590b14975424930fd22292f))
+
+
+
+
+
 ## [0.2.9](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/lib-api@0.2.8...@gtms/lib-api@0.2.9) (2020-05-23)
 
 **Note:** Version bump only for package @gtms/lib-api