+++ packages/service-auth/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.16.0](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/service-auth@0.15.1...@gtms/service-auth@0.16.0) (2020-06-07)
+
+
+### Features
+
+* GEOT-270 get user endpoint ([#102](https://github.com/mariusz-kabala/gtms-backend/issues/102)) ([4e3e81b](https://github.com/mariusz-kabala/gtms-backend/commit/4e3e81b32abeae6dd0cb8feb840c704394476e56))
+
+
+
+
+
+## [0.15.1](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/service-auth@0.15.0...@gtms/service-auth@0.15.1) (2020-06-07)
+
+**Note:** Version bump only for package @gtms/service-auth
+
+
+
+
+
+# [0.15.0](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/service-auth@0.14.3...@gtms/service-auth@0.15.0) (2020-05-26)
+
+
+### Features
+
+* GEOT-266: add favs endpoints ([#98](https://github.com/mariusz-kabala/gtms-backend/issues/98)) ([b296aff](https://github.com/mariusz-kabala/gtms-backend/commit/b296aff2f4b87ad9a590b14975424930fd22292f))
+
+
+
+
+
 ## [0.14.3](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/service-auth@0.14.2...@gtms/service-auth@0.14.3) (2020-05-23)
 
 **Note:** Version bump only for package @gtms/service-auth