+++ packages/worker-tags/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.13](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/worker-tags@0.2.12...@gtms/worker-tags@0.2.13) (2020-06-07)
+
+**Note:** Version bump only for package @gtms/worker-tags
+
+
+
+
+
+## [0.2.12](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/worker-tags@0.2.11...@gtms/worker-tags@0.2.12) (2020-06-07)
+
+**Note:** Version bump only for package @gtms/worker-tags
+
+
+
+
+
+## [0.2.11](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/worker-tags@0.2.10...@gtms/worker-tags@0.2.11) (2020-05-26)
+
+**Note:** Version bump only for package @gtms/worker-tags
+
+
+
+
+
 ## [0.2.10](https://github.com/mariusz-kabala/gtms-backend/compare/@gtms/worker-tags@0.2.9...@gtms/worker-tags@0.2.10) (2020-05-23)
 
 **Note:** Version bump only for package @gtms/worker-tags